use hook for stick to bottom
